Tela de login alterado. Adicionado a mascara de preço em todas pages e listas que possuiam "valor" e/ou "salario". Alterado de "Float" para "Decimal" os metodos das Models que possuiam "Valor" e/ou "Salario". Retirado o codigo obsoleto "Mode=TwoWay" de todas as pages.

+++ Carro/Carro/Carro/Models/Funcionario.cs
@@ -13,7 +13,7 @@ namespace Carro.Models
         [PrimaryKey, AutoIncrement]
         public long? Id { get; set; }
 
-        public float Salario { get; set; }
+        public decimal Salario { get; set; }
 
         public string Funcao { get; set; }
