+++ Carro/Carro/Carro/Models/PerdaProduto.cs
@@ -20,7 +20,7 @@ namespace Carro.Models
 
         public string Nome { get; set; }
 
-        public float Preco { get; set; }
+        public decimal Preco { get; set; }
 
         public int Quantidade { get; set; }
